native: two more real bodies that a return-0 stub was standing in front of
Both were found by scripts/native/check_stub_shadow.py's second arm, which asks
a question the duplicate-definition test cannot: is a stub the ONLY definition
of a symbol that a real object *references*? 121 of the 145 solo stubs are in
that state; almost all are Xbox SDK / Kinect / Bink entry points that correctly
have no body. Two were not.

ValidateThreadId
  src/system/utl/MakeString.cpp had an HX_NATIVE body all along, spelled
      bool ValidateThreadId(DWORD)
  while os/OSFuncs.h declares `bool ValidateThreadId(unsigned long)`. On the
  Xbox toolchain those are the same type; the native xdk shim makes DWORD a
  32-bit `unsigned int`, so the file *defined* _Z16ValidateThreadIdj and
  *referenced* _Z16ValidateThreadIdm at the same time -- visible in one `nm`:

      MakeString.cpp.o : T ValidateThreadId(unsigned int)
      MakeString.cpp.o : U ValidateThreadId(unsigned long)
      MemMgr.cpp.o     : U ValidateThreadId(unsigned long)

  The `unsigned long` overload was satisfied by the weak stub _stub_fn_61, which
  returns 0. Both call sites are `if (!ValidateThreadId(*threadIdSlot))`, so the
  predicate was permanently "this thread is dead" and the first slot of the
  per-thread table was reclaimed on every miss -- in MemMgr::GetMemStack that
  means handing a live thread's MemStack to another thread, past two
  MILO_ASSERTs that are non-fatal on native.

  before: ef0010 lea gStubTraceEnabled(%rip),%rax ... xor %eax,%eax ; ret
  after : dd0c50 push %rbp ; mov %rsp,%rbp ; mov $0x1,%al ; ret

  Fixed inside the existing #ifdef HX_NATIVE, so the PPC side cannot see it.

DspAllocate
  The body dc3-native actually links comes from the shared engine
  (milo-native-engine/src/platform/SynthCommon_Stub.cpp), not from this repo's
  copy -- native/CMakeLists.txt REMOVE_ITEMs the DC3 copy for engine-linking
  targets. Both declared the third parameter void*, which mangles differently
  from the IXAudioBatchAllocator* that synth/Common_Xbox.h declares and that
  DelayEffect.cpp / FlangerEffect.cpp reference, so the real body was dead and
  _stub_fn_8 won. DelayEffect::mBuffer and FlangerEffect::mDelayBuffers[] are
  not in their ctors' init lists, so those effects ran on an indeterminate
  pointer and ~DelayEffect passed it to delete[].

  before: DelayEffect::DelayEffect ... jmp eefef0 <DspAllocate...>  (xor eax,eax; ret)
  after : DelayEffect::DelayEffect ... jmp f755f0 <DspAllocate...>  (call _Znam@plt; mov %rax,(%rbx))

  The root fix landed in milo-native-engine as d62700c (branch
  fix/dspallocate-signature, merged --no-ff to that repo's main). The copy in
  this repo is still compiled for the non-engine targets (dc3-web, export
  tools), so it needed the same correction.

Gate additions in scripts/native/check_stub_shadow.py:
  * --self-test, which pins the stub fingerprint in both directions against
    bodies whose nature is not in question. It immediately earned its keep: the
    first fingerprint scanned objdump's whole listing, so it matched the symbol
    *label line*, and dc3::StubTraceHit -- a real function -- classified as a
    stub. The detector now looks only at instruction lines and requires every
    instruction to be stub-shaped.
  * an unresolved-symbol arm using `ldd -r`, with the accepted set as a
    baseline. `ldd -r` returns exactly what relinking with
    --unresolved-symbols=report-all returns (both give 31 for dc3-native) and
    takes seconds rather than a full relink.

Decomp impact: none. scripts/measure_progress.sh --current-dir on this worktree
reports overall fuzzy 53.99% -> 53.99% (+0.00%), 0 subsystems changed,
+0 matched functions.

+++ src/system/utl/MakeString.cpp
@@ -40,7 +40,18 @@ void TerminateMakeString() {}
 bool MakeStringInitted() { return gLock != nullptr; }
 
 #ifdef HX_NATIVE
-bool ValidateThreadId(DWORD) {
+// The parameter type must be spelled `unsigned long`, exactly as os/OSFuncs.h
+// declares it, NOT `DWORD`. On the Xbox toolchain those are the same type, but
+// the native xdk shim makes DWORD a 32-bit `unsigned int`, so writing DWORD here
+// defined _Z16ValidateThreadIdj while every caller referenced
+// _Z16ValidateThreadIdm -- `nm` on MakeString.cpp.o showed the file *defining*
+// ValidateThreadId(unsigned int) and *referencing* ValidateThreadId(unsigned
+// long) at the same time. The unresolved `unsigned long` overload was then
+// satisfied by the weak stub _stub_fn_61 in engine_stubs_generated.cpp, which
+// returns 0, so `!ValidateThreadId(...)` was permanently true and both callers
+// (this file's per-thread buffer table and MemMgr::GetMemStack) immediately
+// reclaimed slot 0 from a live thread on every miss.
+bool ValidateThreadId(unsigned long) {
     return true; // Thread validation not meaningful on native
 }
 #else
